# swagger.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "PROJECT-XfARM-API",
		"description": "Api para consumo de dados pelo sistema de gestão de fazendas, PROJECT XfARM",
		"contact": {
			"email": "ayvonamorim@gmail.com"
		},
		"version": "0.10"
	},
	"servers": [
		{
			"url": "https://localhost:3000"
		}
	],
	"tags": [
		{
			"name": "client",
			"description": "Requisições que envolvem dados de clientes"
		}
	],
	"paths": {
		"/client": {
			"post": {
				"tags": ["client"],
				"summary": "Cria um Cliente",
				"description": "Cria um novo id de cliente no sistema",
				"operationId": "CreateClient",
				"requestBody": {
					"description": "Cria um novo cadastro de cliente no BD",
					"content": {
						"application/json": {
							"schema":{
                                "$ref":"#/componentes/schemas/Cliente"
                            }
						}
					}
				}
			}
		},
        "/user":{
            "post": {
                "tags": ["user"],
                "summary": "Cria um novo usuário no sistema",
				"operationId": "CreateUser",
                "requestBody": {
					"description": "Cria um novo cadastro de Usuário no BD",
					"content": {
						"application/json": {
							"schema":{
                                "$ref":"#/componentes/schemas/User"
                            }
						}
					}
				}
            }
        }
	},
    "componentes": {
        "schemas": {
            "Cliente": {
                "type": "object",
                "properties": {
                    "name": {
                        "type":"string"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "nome": {
                        "type": "string",
                        "required": true,
                        "trim": true
                    },
                    "sobrenome": {
                        "type": "string",
                        "required": true,
                        "trim": true
                    },
                    "login": {
                        "type": "string",
                        "required": true,
                        "unique": true
                    },
                    "email": {
                        "type": "string",
                        "required": true,
                        "unique": true,
                        "lowercase": true
                    },
                    "senha": {
                        "type": "string",
                        "required": true,
                        "select": false
                    },
                    "client_id": {
                        "type": "mongoose.Schema.Types.ObjectId",
                        "required": true,
                        "ref": "Clientes"
                    },
                    "AdmAccess": {
                        "type": "Boolean"
                    },
                    "ActiveFarms": {
                        "type": "Array",
                        "required": false
                    },
                    "ImgPerfil": {
                        "type": "String",
                        "required": false
                    },
                    "refreshToken": {
                        "type": "string"
                    }
                }
            }
        }
    }
}